[IMP] web: enable specifying a view for a dialog
I want to show a dialog that uses a different view from the default. At present it is possible to specify a searchViewId but not a viewId.

For a custom model, the developer has the ability to create a first view which becomes the default and is used by the dialog. But there seems to be no way of specifying a view for an existing model.